Provide fullpath to windows directory on REPO_ROOT
