+++ packages/osd-cross-platform/src/index.ts
@@ -4,4 +4,3 @@
  */
 
 export * from './path';
-export * from './process';
